Rename file to CosmosAccounts.js
